#1 modalのコンポーネント化
